Refactor ManageCoursesView.tsx and SectionQuizView.tsx
